requirements.txt
azure-cognitiveservices-speech==1.17.0
beautifulsoup4==4.9.3
certifi==2020.12.5
chardet==4.0.0
click==8.0.1
colorama==0.4.4
comtypes==1.1.10
cssselect==1.1.0
django-os-geocoder==0.1.6
docopt==0.6.2
feedfinder2==0.0.4
feedparser==6.0.2
filelock==3.0.12
geographiclib==1.50
geograpy3 @ git+git://github.com/jmbielec/geograpy3.git@21b85924f262c7c36f84ac15fff814001746d547
geopy==2.1.0
idna==2.10
isodate==0.6.0
jellyfish==0.8.2
jieba3k==0.35.1
joblib==1.0.1
Js2Py==0.71
lxml==4.6.3
newspaper3k==0.2.8
nltk==3.6.2
numpy==1.20.3
packaging==20.9
Pillow==8.2.0
pipwin==0.5.1
pocketsphinx @ file:///C:/Users/sasha/pipwin/pocketsphinx-0.1.15-cp39-cp39-win_amd64.whl
progressbar==2.5
PyAudio @ file:///C:/Users/sasha/pipwin/PyAudio-0.2.11-cp39-cp39-win_amd64.whl
pycountry==20.7.3
pyjsparser==2.7.1
pylodstorage==0.0.29
pyparsing==2.4.7
pypiwin32==223
PyPrind==2.11.3
pySmartDL==1.3.4
python-dateutil==2.8.1
pyttsx3==2.90
pytz==2021.1
pywin32==300
PyYAML==5.4.1
rdflib==5.0.0
regex==2021.4.4
requests==2.25.1
requests-file==1.5.1
sgmllib3k==1.0.0
six==1.16.0
soupsieve==2.2.1
South==1.0.2
SPARQLWrapper==1.8.5
SpeechRecognition==3.8.1
timezonefinder==5.2.0
tinysegmenter==0.3
tldextract==3.1.0
tqdm==4.61.0
tzlocal==2.1
urllib3==1.26.4
